Refactor dependencies, messaging setup, and permission checks
Updated Maven dependencies to replace unused and redundant libraries, added `hibernate-spatial`, and replaced `quarkus-hibernate-orm` with updated versions. Standardized Kafka channel names in `DeviceStreamService`, removing redundant slashes. Enhanced permission checks in `PermissionCheckers` to align with updated group naming conventions. Refined `application.yml` to enable Quarkus dev services, update JDBC config defaults, and simplify Keycloak configurations.

+++ utils/src/main/java/co/blueguardian/cerebralstratum/utils/uuid/UUIDv5Generator.java
@@ -38,7 +38,7 @@ public class UUIDv5Generator {
             hash[6] &= 0x0f; // clear version
             hash[6] |= 0x50; // set to version 5
             hash[8] &= 0x3f; // clear variant
-            hash[8] |= 0x80; // set to IETF variant
+            hash[8] |= (byte) 0x80; // set to IETF variant
 
             // Convert hash to UUID
             return toUUID(hash);
